refactor(config): separate agent runtime settings
