+++ backend/internal/config/constants/minimax.go
@@ -1,22 +1,8 @@
 package constants
 
-import (
-	_ "embed"
-	"time"
-)
+import _ "embed"
 
 const (
-	CodexHarnessBinary           = "codex"
-	CodexHarnessPackage          = "@openai/codex"
-	CodexHarnessVersionPin       = "CODEX_CLI_VERSION"
-	CodexHarnessVersionFlag      = "--version"
-	CodexHarnessInstallTimeout   = 5 * time.Minute
-	CodexHarnessWaitTimeout      = 2 * time.Minute
-	CodexHarnessInterruptTimeout = 10 * time.Second
-	CodexHarnessAppServer        = "app-server"
-	CodexHarnessBrowserCommand   = `mcp_servers.browser.command="npx"`
-	CodexHarnessBrowserArgs      = `mcp_servers.browser.args=["@playwright/mcp","--cdp-endpoint","http://127.0.0.1:9222","--caps=vision"]`
-
 	MiniMaxLabel                        = "MiniMax"
 	MiniMaxModel                        = "MiniMax-M3"
 	MiniMaxModelDescription             = "MiniMax M3 with a 1,000,000-token context window"
